Use `luigi.MonthParameter` and encode last month in final output
It took me a while to figure out that `luigi.MonthParameter` exists, which is
exactly what I needed and worked around with by using separate date and year
parameters.

Also added a `last_month` parameter to the final output to force the output to
be recalculated if a new dataset becomes available.