Use `luigi.MonthParameter` and `datetime.date` wherever possible
Previously I encoded months in Luigi with year and month IntParameters, but now
I found that there is MonthParameter so I decided to use that. Also changed the
API of my utility functions to takes months as `datetime.date`. I also replaced
some of the uses of `pd.Timestamp` with pure datetime objects, in order to
prepare a potential transition from Pandas to Polars for example